fixed warnings
removed asserts from the bone code
fixed converstion warnings

+++ src/engine_open_gl/rendering/model_animation/animation.cpp
@@ -8,8 +8,10 @@ namespace model_animation {
 		const aiScene* scene = importer.ReadFile(animationPath, aiProcess_Triangulate);
 		assert(scene && scene->mRootNode);
 		auto animation = scene->mAnimations[0];
-		m_Duration = animation->mDuration;
-		m_TicksPerSecond = animation->mTicksPerSecond;
+		//change the level of prececion for the engine
+		m_Duration = static_cast<float>(animation->mDuration);
+		m_TicksPerSecond = static_cast<float>(animation->mTicksPerSecond);
+
 		aiMatrix4x4 globalTransformation = scene->mRootNode->mTransformation;
 		globalTransformation = globalTransformation.Inverse();
 		ReadHeirarchyData(m_RootNode, scene->mRootNode);
@@ -83,7 +85,7 @@ namespace model_animation {
 		dest.transformation = AssimpGLMHelpers::ConvertMatrixToGLMFormat(src->mTransformation);
 		dest.childrenCount = src->mNumChildren;
 
-		for (int i = 0; i < src->mNumChildren; i++)
+		for (unsigned int i = 0; i < src->mNumChildren; i++)
 		{
 			AssimpNodeData newData;
 			ReadHeirarchyData(newData, src->mChildren[i]);
