+++ src/engine_open_gl/rendering/model_animation/animation.h
@@ -39,7 +39,7 @@ namespace model_animation {
 		void ReadHeirarchyData(AssimpNodeData& dest, const aiNode* src);
 
 		float m_Duration;
-		int m_TicksPerSecond;
+		float m_TicksPerSecond;
 		std::vector<Bone> m_Bones;
 		AssimpNodeData m_RootNode;
 		std::map<std::string, BoneInfo> m_BoneInfoMap;
